HS-95-07: configuration through the desk — settings edit in place
SettingsCore, ProfilesCore, CadenceCore, and SetupCore join the pattern
(Commands landed in 04): flat routes are thin wrappers with their exact
original copy; the desk hosts the cores as windows. The shelf's
Integrations entry is a surface ALIAS — it opens the Settings window
scoped to integrations with the scope chip; configure-integration from
the inspector scopes to the specific integration. The inspector's edit
affordances and the Pullout's capability row open scoped windows in
place; FirstWords' Setup/Runs on go through the shell; the dictation
core's Setup button now resolves in-world too.

Proof (evidence-story-07.md + assets/config-1440.png): suite 258/258
(FirstWords test updated to the dispatcher mechanically); the config walk
on the production bundle — the chrome menu opens Settings in-world, a
REAL presence toggle round-trips through /api/settings and survives
reload, Runs on/Cadence/Integrations open as cascaded windows, and every
flat configuration route still answers for deep links.

PMO-Story: HS-95-07
PMO-Contract-Digest: sha256:03955460d16affec240556db7733c17373356e818bb3c0216f4bcbf7d8b5e6c3
